+++ inc/blueberry-transcoder.h
@@ -54,11 +54,11 @@ typedef struct {
 
 /**
  * this is really an index into the buffer, measured in bytes, although it is assumed to be linear,
- * even if the buffer wraps partway through the packet.
+ * even if the buffer wraps part way through the packet.
  * If this value is i and the packet length is n, then 0 <= i < n.
  * An official invalid value for a block is 0xffff, or BB_INVALID_BLOCK
  */
-typedef uint32_t BbBlock;
+typedef uint16_t BbBlock;
 //typedef uint32_t BbArray;//ditto
 
 //*******************************************************************************************
@@ -73,83 +73,83 @@ typedef uint32_t BbBlock;
 /**
  *  gets an 8-bit, unsigned integer from the specified block
  */
-uint8_t getBbUint8(Bb* buf, BbBlock p, uint32_t i);
+uint8_t getBbUint8(Bb* buf, BbBlock p, uint16_t i);
 
 /**
  * sets an 8-bit, unsigned integer in the specified block
  */
-void setBbUint8(Bb* buf, BbBlock p, uint32_t i, uint8_t v);
+void setBbUint8(Bb* buf, BbBlock p, uint16_t i, uint8_t v);
 
 /**
  * gets an 8-bit, signed integer from thespecified block
  */
-int8_t getBbInt8(Bb* buf, BbBlock p, uint32_t i);
+int8_t getBbInt8(Bb* buf, BbBlock p, uint16_t i);
 
 /**
  * sets an 8-bit signed integer in the specified block
  */
-void setBbInt8(Bb* buf, BbBlock p, uint32_t i, int8_t v);
+void setBbInt8(Bb* buf, BbBlock p, uint16_t i, int8_t v);
 
 
 /**
  *  gets a 16-bit, unsigned integer from the specified block
  */
-uint16_t getBbUint16(Bb* buf, BbBlock p, uint32_t i);
+uint16_t getBbUint16(Bb* buf, BbBlock p, uint16_t i);
 
 /**
  * sets a 16-bit, unsigned integer in the specified block
  */
-void setBbUint16(Bb* buf, BbBlock p, uint32_t i, uint16_t v);
+void setBbUint16(Bb* buf, BbBlock p, uint16_t i, uint16_t v);
 
 /**
  * gets a 16-bit, signed integer from thespecified block
  */
-int16_t getBbInt16(Bb* buf, BbBlock p, uint32_t i);
+int16_t getBbInt16(Bb* buf, BbBlock p, uint16_t i);
 
 /**
  * sets a 16-bit signed integer in the specified block
  */
-void setBbInt16(Bb* buf, BbBlock p, uint32_t i, int16_t v);
+void setBbInt16(Bb* buf, BbBlock p, uint16_t i, int16_t v);
 
 /**
  *  gets a 32-bit, unsigned integer from the specified block
  */
-uint32_t getBbUint32(Bb* buf, BbBlock p, uint32_t i);
+uint32_t getBbUint32(Bb* buf, BbBlock p, uint16_t i);
 
 /**
  * sets a 32-bit, unsigned integer in the specified block
  */
-void setBbUint32(Bb* buf, BbBlock p, uint32_t i, uint32_t v);
+void setBbUint32(Bb* buf, BbBlock p, uint16_t i, uint32_t v);
 
 /**
  * gets a 32-bit, signed integer from thespecified block
  */
-int32_t getBbInt32(Bb* buf, BbBlock p, uint32_t i);
+int32_t getBbInt32(Bb* buf, BbBlock p, uint16_t i);
 
 /**
  * sets a 32-bit signed integer in the specified block
  */
-void setBbInt32(Bb* buf, BbBlock p, uint32_t i, int32_t v);
+void setBbInt32(Bb* buf, BbBlock p, uint16_t i, int32_t v);
 
 /**
  *  gets a 32-bit, floating point value from the specified block
  */
-float getBbFloat32(Bb* buf, BbBlock p, uint32_t i);
+float getBbFloat32(Bb* buf, BbBlock p, uint16_t i);
 
 /**
  * sets a a 32-bit, floating point value in the specified block
  */
-void setBbFloat32(Bb* buf, BbBlock p, uint32_t i, float v);
+void setBbFloat32(Bb* buf, BbBlock p, uint16_t i, float v);
 
 /**
  * extracts a boolean from the specified block
  */
-bool getBbBool(Bb* buf, BbBlock p, uint32_t i, uint32_t bitNum);
+bool getBbBool(Bb* buf, BbBlock p, uint16_t i, uint32_t bitNum);
 
 /**
  * sets a boolean in a specified block
  */
-void setBbBool(Bb* buf, BbBlock p, uint32_t i, uint32_t bitNum, bool v);
+void setBbBool(Bb* buf, BbBlock p, uint16_t i, uint32_t bitNum, bool v);
 
 /**
  * converts a linear index to a circular one
